refactor: extract gas estimation utilities
