+++ apps/web/src/features/counterfactual/ActivateAccountFlow.tsx
@@ -11,7 +11,8 @@ import { selectUndeployedSafe } from '@/features/counterfactual/store/undeployed
 import { CF_TX_GROUP_KEY, extractCounterfactualSafeSetup, isPredictedSafeProps } from '@/features/counterfactual/utils'
 import useChainId from '@/hooks/useChainId'
 import { useCurrentChain } from '@/hooks/useChains'
-import useGasPrice, { getTotalFeeFormatted } from '@/hooks/useGasPrice'
+import useGasPrice from '@/hooks/useGasPrice'
+import { getTotalFeeFormatted } from '@safe-global/utils/utils/gasPrice'
 import { useLeastRemainingRelays } from '@/hooks/useRemainingRelays'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import useWalletCanPay from '@/hooks/useWalletCanPay'
